+++ sap/__init__.py
@@ -1,0 +1,2 @@
+''' sap namespace '''
+__path__ = __import__('pkgutil').extend_path(__path__, __name__)