+++ tests/unit/test_init.py
@@ -1,12 +1,12 @@
 """ Tests for cf_logging.init """
 import logging
 import pytest
-import cf_logging
+from sap import cf_logging
 
-from cf_logging.core.constants import REQUEST_KEY, RESPONSE_KEY
-from cf_logging.core.framework import Framework
-from cf_logging.record.request_log_record import RequestWebRecord
-from cf_logging.record.simple_log_record import SimpleLogRecord
+from sap.cf_logging.core.constants import REQUEST_KEY, RESPONSE_KEY
+from sap.cf_logging.core.framework import Framework
+from sap.cf_logging.record.request_log_record import RequestWebRecord
+from sap.cf_logging.record.simple_log_record import SimpleLogRecord
 
 # pylint: disable=protected-access