+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+
+## [Unreleased]
+ - Introduced sap namespace
+
 ## 3.0.1 - 2017-12-11
 
 ### Changed